+++ bot/main.py
@@ -1,12 +1,19 @@
+"""
+Start program
+"""
+
+import logging
+import os
+
 from resident_control import Auth, HandleEntryVisitor
 from register import Register
 from register_visitor import RegisterVisitor
 from feedback import Feedback
+from notify import NotifyAdmin
 from visit import Visit
 from settings import *
 from telegram.ext import Updater, CommandHandler, MessageHandler, ConversationHandler, Filters
-import logging
-import os
+from telegram.ext import CallbackQueryHandler
 
 # Remove logs from APIs
 logging.getLogger("telegram").setLevel(API_LOG_LEVEL)
@@ -25,13 +32,27 @@ file_handler.setFormatter(f_format)
 logger.addHandler(file_handler)
 
 def start(update, context):
-    logger.info("Introducing the bot")
-    update.message.reply_text('Olá, bem vindo(a) ao bot do Alohomora!')
-    update.message.reply_text('Caso deseje fazer uma solicitação de visita a algum morador, digite /visitar')
-    update.message.reply_text('Digite /cadastrar para fazer o cadastro de um morador')
-    update.message.reply_text('Digite /autorizar para autorizar entrada de algum visitante')
-    update.message.reply_text('Para dar um feedback pro nosso serviço, digite /feedback')
-
+    """
+    Start interaction
+    """
+    logger.info(
+        "Introducing the bot"
+    )
+    update.message.reply_text(
+        'Olá, bem vindo(a) ao bot do Alohomora!'
+    )
+    update.message.reply_text(
+        'Caso deseje fazer uma solicitação de visita a algum morador, digite /visitar'
+    )
+    update.message.reply_text(
+        'Digite /cadastrar para fazer o cadastro de um morador'
+    )
+    update.message.reply_text(
+        'Digite /autorizar para autorizar entrada de algum visitante'
+    )
+    update.message.reply_text(
+        'Para dar um feedback pro nosso serviço, digite /feedback'
+    )
 
 if __name__ == '__main__':
 
@@ -109,6 +130,10 @@ if __name__ == '__main__':
 
         fallbacks=[CommandHandler('cancelar', Feedback.end)]
         ))
+
+    # Admin
+    dp.add_handler(CallbackQueryHandler(NotifyAdmin.approved, pattern='app'))
+    dp.add_handler(CallbackQueryHandler(NotifyAdmin.rejected, pattern='rej'))
 
 
     updater.start_polling()