+++ bot/register.py
@@ -1,5 +1,20 @@
+<<<<<<< HEAD
 import librosa
+=======
+"""
+Register a user interaction
+"""
+import json
+import logging
+import os
+import subprocess
+import numpy
+import requests
+
+>>>>>>> a730943efcbdfe2e38b4c9b1e9cc91a63e500ca6
 from checks import CheckResident, CheckCondo
+from db.schema import create_resident
+from notify import NotifyAdmin
 from python_speech_features import mfcc
 from scipy.io.wavfile import read
 from settings import LOG_NAME
@@ -8,20 +23,21 @@ from settings import PATH, CATCH_AUDIO_SPEAKING_NAME, CONFIRM_AUDIO_SPEAKING_NAM
 from telegram import KeyboardButton, ReplyKeyboardMarkup
 from telegram.ext import ConversationHandler
 from validator import ValidateForm
-import json
-import logging
-import numpy
-import os
-import requests
-import subprocess
+
 
 logger = logging.getLogger(LOG_NAME)
 
 chat = {}
 
 class Register:
+    """
+    Register a resident
+    """
 
     def index(update, context):
+        """
+        Start the conversation
+        """
         logger.info("Introducing registration session")
         chat_id = update.message.chat_id
 
@@ -36,6 +52,9 @@ class Register:
         return NAME
 
     def name(update, context):
+        """
+        Get name of a resident
+        """
         chat_id = update.message.chat_id
         name = update.message.text
 
@@ -46,8 +65,12 @@ class Register:
         logger.debug(f"'name': '{chat[chat_id]['name']}'")
 
         contact_keyboard = KeyboardButton('Enviar meu número de telefone', request_contact=True)
-        custom_keyboard = [[ contact_keyboard ]]
-        reply_markup = ReplyKeyboardMarkup(custom_keyboard, one_time_keyboard=True, resize_keyboard=True)
+        custom_keyboard = [[contact_keyboard]]
+        reply_markup = ReplyKeyboardMarkup(
+            custom_keyboard,
+            one_time_keyboard=True,
+            resize_keyboard=True
+        )
 
         update.message.reply_text('Telefone:', reply_markup=reply_markup)
         logger.info("Asking for phone")
@@ -55,6 +78,9 @@ class Register:
         return PHONE
 
     def phone(update, context):
+        """
+        Get phone information
+        """
         chat_id = update.message.chat_id
         phone = update.message.text
         contact = update.effective_message.contact
@@ -73,6 +99,10 @@ class Register:
         return EMAIL
 
     def email(update, context):
+        """
+        Get email information
+        """
+
         chat_id = update.message.chat_id
         email = update.message.text
 
@@ -97,6 +127,10 @@ class Register:
         return CPF
 
     def cpf(update, context):
+        """
+        Get cpf information
+        """
+
         chat_id = update.message.chat_id
         cpf = update.message.text
 
@@ -125,6 +159,10 @@ class Register:
 
 
     def block(update, context):
+        """
+        Get block information
+        """
+
         chat_id = update.message.chat_id
         block = update.message.text
 
@@ -149,6 +187,10 @@ class Register:
         return APARTMENT
 
     def apartment(update, context):
+        """
+        Get apartment information
+        """
+
         chat_id = update.message.chat_id
         apartment = update.message.text
 
@@ -260,7 +302,7 @@ Escute o audio gravado e verifique se:
         # Repeat and confirm buttons
         repeat_keyboard = KeyboardButton('Repetir')
         confirm_keyboard = KeyboardButton('Confirmar')
-        keyboard = [[repeat_keyboard],[confirm_keyboard]]
+        keyboard = [[repeat_keyboard], [confirm_keyboard]]
         choice = ReplyKeyboardMarkup(keyboard, resize_keyboard=True, one_time_keyboard=True)
         update.message.reply_text('Escute o seu áudio e confirme se está com boa qualidade', reply_markup=choice)
 
@@ -269,6 +311,10 @@ Escute o audio gravado e verifique se:
         return REPEAT_VOICE
 
     def repeat_voice(update, context):
+        """
+        Repeate voice interaction
+        """
+
         chat_id = update.message.chat_id
         choice = update.message.text
 
@@ -297,22 +343,33 @@ Escute o audio gravado e verifique se:
         os.remove(audio_file_path)
         os.remove(wav_audio_file_path)
 
+        logger.debug(f"data['{chat_id}']: {chat[chat_id]}")
         response = Register.register_resident(chat_id)
 
         if(response.status_code == 200 and 'errors' not in response.json().keys()):
-            logger.info("resident registered in database")
-            update.message.reply_text('Morador cadastrado no sistema!')
+            logger.info("Resident registered in database")
+            update.message.reply_text("Cadastro feito!\nAguarde aprovação dos administradores.")
+
+            create_resident(
+                    cpf=chat[chat_id]['cpf'],
+                    block=chat[chat_id]['block'],
+                    apartment=chat[chat_id]['apartment'],
+                    chat_id=chat_id
+                    )
+            NotifyAdmin.send_message(context, chat[chat_id])
         else:
             logger.error("Registration failed")
             update.message.reply_text('Falha ao cadastrar no sistema!')
-
-        # logger.debug(f"data['{chat_id}']: {chat[chat_id]}")
 
         chat[chat_id] = {}
 
         return ConversationHandler.END
 
     def end(update, context):
+        """
+        Cancel interaction
+        """
+
         logger.info("Canceling registration")
         chat_id = update.message.chat_id
 
@@ -324,6 +381,10 @@ Escute o audio gravado e verifique se:
         return ConversationHandler.END
 
     def register_resident(chat_id):
+        """
+        Register a resident
+        """
+
         logger.info("Registering resident")
         query = """
         mutation createResident(